Optionally use only MAM for MUC history
Adds a setting to the "expert options" to exclusively use MAM for
history retrieval in MUCs (when they support MAM). Local history is
thus deleted upon rejoining.

Conversations used to only retrieve messages older than the oldest
locally known message using MAM once joined. Thus there might be
messages missing between the last disconnect and the new connect
because these messages are only retrieved when joining and are limited.

This option can be turned on if you are usually not interested in
older history and are more concerned about the recent history of the
MUC (useful for very active MUCs).